% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_generator.R
\name{mcmc_generator}
\alias{mcmc_generator}
\title{Generate some MCMC traces for BotB poster: "Now you see it, now you don't"}
\usage{
mcmc_generator(max_mixtures, burn_in, samples, extra_sample)
}
\arguments{
\item{max_mixtures}{iteger, the maximum number of mixtures to generate}

\item{burn_in}{integer size of the burn-in}

\item{samples}{integer size of the sample run}

\item{extra_sample}{integer size of the extra run at the end, which we use
the samples for}
}
\value{
a list containing three items - max_mixtures - the maximum number
    of mixture models to fit. sample_eyes - the samples of the eyes model
    extra_eyes - the samples of the extra samples we take at the end
}
\description{
Generate some MCMC traces for BotB poster: "Now you see it, now you don't"
}
\examples{

\dontrun{

lawful_good <- mcmc_generator(max_mixtures = 2,
                              burn_in = 10000,
                              samples = 50000,
                              extra_sample = 10000)

chaotic_evil <- mcmc_generator(max_mixtures = 48,
                               burn_in = 10000,
                               samples = 50000,
                               extra_sample = 10000)


}
}
